A bunch of typos Bench: 2191122

+++ src/eval/endgame.cpp
@@ -10,7 +10,7 @@
 
 // Returns an integer between -width ~ width cp as a draw score
 // This makes the engine prefer positions with higher mobility (more likely to get higher scores)
-int8_t endgame_noise(unsigned int seed, uint8_t width) {
+int16_t endgame_noise(unsigned int seed, uint8_t width) {
     std::mt19937_64 mt(seed);  // 64-bit Mersenne Twister
     std::uniform_int_distribution<int> distribution(-width, width);
     return distribution(mt);
